requirements.txt
asgiref==3.4.1
Django==3.2.13
djangorestframework==3.13.1
pkg-resources==0.0.0
pytz==2022.1
sqlparse==0.4.2
typing-extensions==4.1.1
psycopg2
PyJWT==v1.7.1
djangorestframework-simplejwt
